+++ src/pip/_internal/req/constructors.py
@@ -23,6 +23,7 @@ from pip._vendor.pkg_resources import RequirementParseError, parse_requirements
 from pip._internal.exceptions import InstallationError
 from pip._internal.models.index import PyPI, TestPyPI
 from pip._internal.models.link import Link
+from pip._internal.models.wheel import Wheel
 from pip._internal.pyproject import make_pyproject_path
 from pip._internal.req.req_install import InstallRequirement
 from pip._internal.utils.filetypes import ARCHIVE_EXTENSIONS
@@ -30,7 +31,6 @@ from pip._internal.utils.misc import is_installable_dir, splitext
 from pip._internal.utils.typing import MYPY_CHECK_RUNNING
 from pip._internal.utils.urls import path_to_url
 from pip._internal.vcs import is_url, vcs
-from pip._internal.wheel import Wheel
 
 if MYPY_CHECK_RUNNING:
     from typing import (
