+++ tests/unit/test_wheel.py
@@ -13,12 +13,12 @@ from pip._internal.commands.wheel import WheelCommand
 from pip._internal.exceptions import InvalidWheelFilename, UnsupportedWheel
 from pip._internal.locations import get_scheme
 from pip._internal.models.scheme import Scheme
+from pip._internal.models.wheel import Wheel
 from pip._internal.utils.compat import WINDOWS
 from pip._internal.utils.misc import hash_file
 from pip._internal.utils.unpacking import unpack_file
 from pip._internal.wheel import (
     MissingCallableSuffix,
-    Wheel,
     _raise_for_invalid_entrypoint,
 )
 from pip._internal.wheel_builder import get_legacy_build_wheel_path
